Fix a reindenter bug (E123)
The highlighted case was missing "indent_level" worth of indenting from the
adjusted indent; this was because the "indent_level" wasn't being added on,
and also because the list of tokens for the logical line included leading
indent tokens.  Fix both problems and add the case to the test suite.