Don't split a container if it will cause the alignment to go haywire.
If a container is split towards the end of the line, a further message
may come along and claim that the alignment is off. But fixing that
alignment will cause it to go over the maximum line length.

Apply a heuristic that doesn't split a container if it's too late on the
line.