Expand the experimental E501 tests.
This adds "experimental" tests for all of the existing E501 test cases.
It also splits out the experimental tests into their own class.

The good thing is that virtually all tests have the same or better
output as their aggressive counterparts. However, there are a few which
have not-so-pretty output. Those have been marked with FIXMEs.