Move the logic to prevent splitting default initializers and ending delimiters.
This moves the logic that prevents splitting between default
initializers and before an ending delimiter into the ReflowedLines
class. Instead of the external objects trying to guess and where to
split based on next tokens, etc., the ReflowedLines object does it
itself. It knows about all of the items that are already inserted and
can manipulate them much better than an outside class/object.