Penalize short continued lines more than long ones.
If a line ends in \-newline, we check to see if it's splitting up a long
line or a short line. If it's a long line, or the line has comments on
it, then we don't penalize the ranking too much. However, if it's
splitting up lines which don't go over the maximum line length, then we
want to penalize it a lot, because it doesn't look quite as nice as the
non-split line.