w605 fixes for identical tokens
An infinite loop was occuring when two w605 tokens appeared on the same line because the string find() function would only return the position of the first instance to be fixed.  The Token datastrucute (tuple) already contained the start position where the fix was needed, so it is used directly instead of doing a search.  Unit tests added that catch the original issue.