autopep8.py: Add ability to specify a column range
This adds the ability to specify a column range
on which to perform fixes. Although multiline
column range fixing is supported, it is not
recommended, since changing columns may result in
differing numbers of columns among lines - which
would make it impossible to automatically adjust
the column range after performing changes.

Closes https://github.com/hhatto/autopep8/issues/227